requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.1.5
django-crispy-forms==1.7.2
django-tinymce==3.3.0
gunicorn==20.1.0
idna==3.1
multidict==4.7.6
Pillow==8.2.0
pytz==2021.1
websockets==8.1
whitenoise==5.2.0
yarl==1.6.3
